vtxo: move the expiry guard tests into service_test.go
expiry_guards_test.go had no source file to mirror. This package names its test
files after the file under test, and both helpers live in service.go, so the
tests belong in service_test.go with the rest of it.

TestCheckUnrolledVtxoExpiry settles it. It covers the function sitting directly
below these two helpers in service.go and is already in service_test.go, so a
separate file for their neighbours would have split the expiry tests across two
places for no reason. They now sit next to it.

arkfee_test.go stays as it is. It mirrors arkfee.go, which is the only source
file in the feemanager package, and there was no existing test file to join.
